requirements.txt
alembic==1.13.2
fastapi==0.112.0
fastapi_utils==0.7.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic[email]==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.32
uvicorn==0.30.5
pytest==8.3.2
pytest-env==1.1.3
